refactor(api/middleware): move inside server dir
