+++ api/server/middleware/validateRegistration.js
