+++ api/server/routes/ask/openAI.js
@@ -5,7 +5,7 @@ const { parseConvo } = require('librechat-data-provider');
 const { getAzureCredentials, abortMessage } = require('../../../utils');
 const { saveMessage, getConvoTitle, saveConvo, getConvo } = require('../../../models');
 const { handleError, sendMessage, createOnProgress } = require('../handlers');
-const requireJwtAuth = require('../../../middleware/requireJwtAuth');
+const requireJwtAuth = require('../../middleware/requireJwtAuth');
 
 const abortControllers = new Map();
