+++ api/server/routes/messages.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const { getMessages } = require('../../models/Message');
-const requireJwtAuth = require('../../middleware/requireJwtAuth');
+const requireJwtAuth = require('../middleware/requireJwtAuth');
 
 router.get('/:conversationId', requireJwtAuth, async (req, res) => {
   const { conversationId } = req.params;
