+++ api/server/routes/plugins.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { getAvailablePluginsController } = require('../controllers/PluginController');
-const requireJwtAuth = require('../../middleware/requireJwtAuth');
+const requireJwtAuth = require('../middleware/requireJwtAuth');
 
 const router = express.Router();
