+++ api/server/routes/user.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const requireJwtAuth = require('../../middleware/requireJwtAuth');
+const requireJwtAuth = require('../middleware/requireJwtAuth');
 const { getUserController, updateUserPluginsController } = require('../controllers/UserController');
 
 const router = express.Router();
